Finally opening a file in the buffer list!
I had to change some things to make this happen. First, I moved the mmap
line to main.c file. This kind of make sense: it's better to let the
main program to manage the memory used with the open files than to keep
all buffers always available in the memory, whithin the core library.